+++ css/basket.js
@@ -26,7 +26,7 @@ function addToBasket(name, price, dose) {
 function addToBasketVariant(selectId, name) {
   const select = document.getElementById(selectId);
   const parts = select.value.split('|');
-  const price = parseInt(parts[0]);
+  const price = parseFloat(parts[0]);
   const dose = parts[2];
   addToBasket(name, price, dose);
 }
